add new function bovigo\assert\assertThat() and make bovigo\assert\assert() an alias of it, fixes #7
Unfortunately, since PHP 7 assert is a keyword which behaves differently
depending on the zend.assertions ini setting. Unfortunately this
also influences user-defined functions when imported into the current
namespace and then called with its short name, as usually done with
bovigo\assert\assert(). The only safe way to get around this is to
not use bovigo\assert\assert() any more and have a different name
for it. As simply renaming it would introduce a major BC break we
simply add a new function, and redirect all calls to the old function
to the new one. This doesn't solve the problem for existing users
but will allow to migrate more easily.